pyproject.toml
[project]
name = "leafy-ils-fast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.4",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.9.0",
    "sqlmodel>=0.0.22",
]
